Move DatePicker, TimePicker and CalendarPanel tests into their own test classes
